Test DGC

Prueba de Di Rienzo, Guzmán y Casanoves para comparaciones múltiples. Implementa un método basado en clusters para identificar grupos de medias no homogéneas.

Usage

dgc_test(y, trt, alpha = 0.05, show_plot = T, console = T, abline_options, ...)

Arguments

y

O bien un modelo (creado con lm() o aov()) o bien un vector numerico con los valores de la variable respuesta para cada unidad.

trt

Si y es un modelo, corresponde a la columna que indica los tratamientos. Si y es un vector, corresponde a un vector de la misma longitud que y con los tratamientos para cada unidad.

alpha

Valor equivalente a 0.05 o 0.01, correspondiente al nivel de significación del test. Por defecto toma el valor 0.05.

show_plot

Valor logico indicando si debe graficarse el dendrograma construido o no.

console

Valor logico indicando si deben imprimirse los resultados en consola o no.

abline_options

Lista con argumentos opcionales para la linea del dendograma.

...

Argumentos opcionales para la funcion plot().

Value

Una lista con tres objetos data.frame:

estadisticas

data.frame con estadisticas resumen segun tratamiento.

grupos

data.frame indicando el grupo al que es asignado cada tratamiento.

parametros

data.frame con los valores utilizados para el test. tratamientos es el numero total de tratamientos, alpha es el nivel de significación utilizado, c es el criterio de corte (indica la altura de la linea horizontal del dendrograma), q es el cuantil 1-alpha de la distribucion de Q (distancia del nodo raiz) bajo la hipotesis nula y SEM es una estimacion del error estandar de la media.

References

Di Rienzo, J.A., Guzmán, A.W., Casanoves, F. (2002): A multiple comparisons method based on the distribution of the root node distance of a binary tree. J. Agr.Biol. Environ. Stat. 7: 1-14.

Author

Santiago Garcia Sanchez

Examples

data("PlantGrowth")
# Utilizando vectores -------------------------------------------------------
pesos <- PlantGrowth$weight
tratamientos <- PlantGrowth$group
dgc_test(y = pesos, trt = tratamientos, show_plot = FALSE)
#>      grupo
#> ctrl     1
#> trt1     1
#> trt2     2
#> Los tratamientos de un mismo grupo no presentan diferencias significativas
# Utilizando un modelo ------------------------------------------------------
modelo <- lm(pesos ~ tratamientos)
dgc_test(y = modelo, trt = "tratamientos", show_plot = FALSE)
#>      grupo
#> ctrl     1
#> trt1     1
#> trt2     2
#> Los tratamientos de un mismo grupo no presentan diferencias significativas