Di Rienzo, Guzman and Casanoves test for multiple comparisons

Di Rienzo, Guzman and Casanoves (DGC) test for multiple comparisons. Implements a cluster-based method for identifying groups of nonhomogeneous means.

Usage

dgc_test(
  y,
  trt,
  alpha = 0.05,
  show_plot = TRUE,
  console = TRUE,
  abline_options,
  ...
)

Arguments

y

Either a model (created with lm() o aov()) or a numerical vector with the values of the response variable for each unit.

trt

If y is a model, it corresponds to the column indicating the treatments. If y is a vector, it's a vector of the same length as y with the treatments for each unit.

alpha

Value equivalent to 0.05 or 0.01, corresponding to the significance level of the test. The default value is 0.05.

show_plot

Logical value indicating whether the constructed dendrogram should be plotted or not.

console

Logical value indicating whether the results should be printed on the console or not.

abline_options

list with optional arguments for the line in the dendrogram.

...

Optional arguments for the plot() function.

Value

A list with three data.frame:

stats

data.frame containing summary statistics by treatment.

groups

data.frame indicating the group to which each treatment is assigned.

parameters

data.frame with the values used for the test. treatments is the total number of treatments, alpha is the significance level used, c is the cut-off criterion for the dendrogram (the height of the horizontal line of the dendrogram), q is the 1-alpha quantile of the distribution of Q (distance from the root node) under the null hypothesis and SEM SEM is an estimate of the standard error of the mean.

References

Di Rienzo, J.A., Guzman, A.W., Casanoves, F. (2002): A multiple comparisons method based on the distribution of the root node distance of a binary tree. J. Agr.Biol. Environ. Stat. 7: 1-14.

Author

Santiago Garcia Sanchez

Examples

data("PlantGrowth")
# Using vectors -------------------------------------------------------
weights <- PlantGrowth$weight
treatments <- PlantGrowth$group
dgc_test(y = weights, trt = treatments, show_plot = FALSE)
#>      group
#> ctrl     1
#> trt1     1
#> trt2     2
#> Treatments within the same group are not significantly different
# Using a model ------------------------------------------------------
model <- lm(weights ~ treatments)
dgc_test(y = model, trt = "treatments", show_plot = FALSE)
#>      group
#> ctrl     1
#> trt1     1
#> trt2     2
#> Treatments within the same group are not significantly different